roslaunch2: prevent exception occurring in silent_remove(path) with path=None

+++ src/roslaunch2/utils.py
@@ -36,6 +36,8 @@ def silent_remove(path):
     :param path: path to file which should be deleted silently
     :return: Returns True if the file was removed, False otherwise.
     """
+    if not path:
+        return False
     try:
         os.remove(path)
         return True
